feat: Moves App component to container dir
fix: Fixes eslint errors
