+++ frontend/src/components/App/App.js
@@ -1,0 +1,10 @@
+import * as React from 'react';
+import Routes from '../Routes/Routes';
+
+const App = () => (
+  <div className="App">
+    <Routes />
+  </div>
+);
+
+export default App;
