+++ frontend/src/components/NavBar/NavBar.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import NavStyles from '../NavBar/NavStyles.module.css';
+import NavStyles from './NavStyles.module.css';
 import logo from '../../assets/robot.jpg';
 
 const NavBar = () => (
   <>
-      <nav className={NavStyles.nav}>
-        <h2 className="logo">
-          <Link to="/">
-            <img src={logo} alt={logo} />
-          </Link>
-          Robot Market
-        </h2>
+    <nav className={NavStyles.nav}>
+      <h2 className="logo">
+        <Link to="/">
+          <img src={logo} alt={logo} />
+        </Link>
+        Robot Market
+      </h2>
       <ul className={NavStyles.navLinks}>
         <Link className={NavStyles.lists} to="/">
           <li>Home</li>
