--- frontend/src/container/App.js
@@ -1,12 +1,0 @@
-import * as React from 'react';
-import Routes from '../components/Routes/Routes';
-
-const App = () => {
-  return (
-    <div className="App">
-      <Routes />
-    </div>
-  );
-};
-
-export default App;
